Refactor: Utiliza o design pattern repository
