+++ src/controllers/usuarios.js
@@ -1,8 +1,10 @@
 const Usuarios = require("../models/usuarios");
 
 module.exports = (app) => {
-  app.get("/usuarios", (req, res, next) => {
-    Usuarios.listar(res, next);
+  app.get("/usuarios", (req, res) => {
+    Usuarios.listar()
+      .then((resultados) => res.json(resultados))
+      .catch((erros) => res.status(400).json(erros));
   });
 
   app.get("/usuarios/:id", (req, res, next) => {
