+++ src/index.test.js
@@ -1,7 +1,7 @@
-const pool = undefined; //require("./infraestrutura/database/conexao");
-const Tabelas = undefined; //require("./infraestrutura/database/tabelas");
-const queries = undefined; //require("./infraestrutura/database/queries");
-const usuariosRepository = undefined; //require("./repositorios/usuario");
+const pool = require("./infraestrutura/database/conexao");
+const Tabelas = require("./infraestrutura/database/tabelas");
+const queries = require("./infraestrutura/database/queries");
+const usuariosRepository = require("./repositorio/usuario");
 const eventosController = undefined; //require("./controllers/eventos");
 const eventosModel = undefined; //require("./models/eventos");
 const tiposVendasController = undefined; //require("./controllers/tiposVendas");
