+++ src/models/usuarios.js
@@ -1,16 +1,10 @@
-const pool = require("../infraestrutura/conexao");
+const pool = require("../infraestrutura/database/conexao");
+const repositorio = require("../repositorio/usuario");
 const fetch = require("node-fetch");
 
 class Usuarios {
-  listar(res, next) {
-    const sql = "SELECT * FROM Usuarios";
-    pool.query(sql, (erro, resultados) => {
-      if (erro) {
-        next(erro);
-      } else {
-        res.status(200).json(resultados);
-      }
-    });
+  listar() {
+    return repositorio.listar();
   }
 
   buscarPorId(id, res, next) {
